requirements.txt
# VulnHawk - Web Application Vulnerability Scanner
# Production Dependencies

# Core Framework
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-wtf>=1.2.0
flask-login>=0.6.3
flask-talisman>=1.1.0
flask-limiter>=3.5.0
flask-socketio>=5.3.6

# Async HTTP Client
aiohttp>=3.9.0
aiofiles>=23.2.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Security
python-dotenv>=1.0.0
cryptography>=41.0.0
bcrypt>=4.1.0
itsdangerous>=2.1.0

# Report Generation
reportlab>=4.0.0
jinja2>=3.1.0
markdown>=3.5.0

# Database
sqlalchemy>=2.0.0

# Validation & Utilities
validators>=0.22.0
colorama>=0.4.6
tqdm>=4.66.0
python-dateutil>=2.8.0
urllib3>=2.1.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# WebSocket
python-socketio>=5.10.0
eventlet>=0.34.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
bandit>=1.7.0
safety>=2.3.0
